requirements.txt
django==2.2
six
django-rest-framework
django-cors-headers
drf_yasg
djangorestframework-api-key
drf-friendly-errors
drf-yasg-examples
mysqlclient
djangorestframework-camel-case
fastecdsa
django-crispy-forms
